refactor(backend): ensure to map weight when returning from db

+++ backend/src/main/java/com/marmouset/workout/external/database/exercise/trained/TrainedExerciseEntity.java
@@ -97,7 +97,7 @@ public class TrainedExerciseEntity extends AbstractEntity {
     return this;
   }
 
-  public float getWeight() {
+  public Integer getWeight() {
     return weight;
   }
